ultimos cambios

+++ src/main/java/com/dh/ClinicMVC/repository/IPacienteRepository.java
@@ -3,5 +3,5 @@ package com.dh.ClinicMVC.repository;
 import com.dh.ClinicMVC.entity.Paciente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface IPacienteRepository extends JpaRepository<Paciente, Long> {
 }
