+++ src/main/java/com/dh/ClinicMVC/repository/ITurnoRepository.java
@@ -3,5 +3,5 @@ package com.dh.ClinicMVC.repository;
 import com.dh.ClinicMVC.entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ITurnoRepository extends JpaRepository<Turno, Integer> {
+public interface ITurnoRepository extends JpaRepository<Turno, Long> {
 }
