+++ src/main/java/com/dh/ClinicMVC/service/IOdontologoService.java
@@ -10,8 +10,8 @@ public interface IOdontologoService {
 
     List<Odontologo> listarTodos();
 
-    Optional<Odontologo> buscarPorId(Integer id);
-    void eliminar(Integer id);
+    Optional<Odontologo> buscarPorId(Long id);
+    void eliminar(Long id);
     void actualizar(Odontologo odontologo);
 
 }
